requirements.txt
# Core scientific stack
numpy
tqdm
matplotlib
pandas
esm
Biopython
h5py
goatools
kaggle

# CUDA-enabled PyTorch (2.2.0, CUDA 12.1)
--extra-index-url https://download.pytorch.org/whl/cu121
torch
torchvision
torchaudio

# ML / Training tools
torchmetrics
pytorch-lightning
wandb

typing-extensions
